+++ src/test/java/nextstep/courses/infrastructure/CourseRepositoryTest.java
@@ -1,6 +1,7 @@
 package nextstep.courses.infrastructure;
 
 import nextstep.courses.domain.course.Course;
+import nextstep.courses.domain.course.CourseRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
